refactor(testing): Extract test-impact core into @n8n/test-impact (no-changelog)
Move the ts-morph-free TIA core (coverage-map, orchestrator, select-e2e) +
their tests out of @n8n/playwright-janitor into a new framework-agnostic
@n8n/test-impact package. DiscoveredSpec is now owned by the package; janitor
re-exports it. Behaviour-preserving move (git mv); 58 package tests + janitor
build/tests green.
