+++ packages/@n8n/test-impact/package.json
@@ -1,15 +1,17 @@
 {
-  "name": "@n8n/playwright-janitor",
+  "name": "@n8n/test-impact",
   "private": true,
   "version": "0.1.0",
-  "description": "Static analysis and architecture enforcement for Playwright test suites",
+  "description": "Test Impact Analysis: coverage-map selection + shard orchestration for n8n CI. Framework-agnostic core.",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
-  "bin": {
-    "janitor": "bin/janitor.mjs"
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "default": "./dist/index.js"
+    }
   },
   "files": [
-    "bin",
     "dist"
   ],
   "scripts": {
@@ -28,28 +30,17 @@
     "typecheck": "tsc --noEmit"
   },
   "keywords": [
-    "playwright",
     "testing",
-    "static-analysis",
-    "architecture",
-    "linting"
+    "test-impact-analysis",
+    "coverage",
+    "ci",
+    "sharding"
   ],
   "license": "MIT",
-  "dependencies": {
-    "@n8n/rules-engine": "workspace:*",
-    "@n8n/test-impact": "workspace:*",
-    "glob": "^10.0.0",
-    "yaml": "catalog:"
-  },
-  "peerDependencies": {
-    "ts-morph": "catalog:"
-  },
   "devDependencies": {
     "@types/node": "catalog:",
     "@vitest/coverage-v8": "catalog:",
     "fast-check": "catalog:",
-    "ts-morph": "catalog:",
-    "tsx": "catalog:",
     "typescript": "catalog:",
     "vitest": "catalog:"
   }
